Read and report memory, disk, and network statistics while checking CPU statistics
Right now, when setting up a performance monitor that reports CPU status at a regular interval, all other statistics (memory, disk, network) are returned ONLY on the first callback. For that reason, this commit does not enhance the 'sysmontap' command, but instead adds a one-shot 'sysstats' command that prints everything.  This approach does, unfortunately, involve creating and tearing down a channel each time we want to read these additional stats.

The call has a mandatory ~1 second delay to include 1-shot CPU stats, as the first reported CPU reading appears to occur with a minimal (or no) sampling interval, often returning "0%" CPU usage.  So we read everything *but* CPU stats from the first stats message, CPU info from the *second* message, then tear down the channel.

We could add memory to sysmontap.  At the moment, this could only be done by building/destroying the channel each time.

This includes a bugfix for closing a SysMonTapService. The Disaptch() callback must not crash or hang, even when called AFTER the underlying channel is closed.  There is an inherent race condition.  Before the fix, SysMonTapService.Close() was closing the messages channel used by its Dispatch() method.  This would lead to random panics when closing the service.  The fix leaves the messages channel open to be garbage collected, and ensures that calls to Dispatch() continue to be serviced after the service is closed(to avoid deadlock or goroutine leaks).